fix build errors (#4)
