+++ src/app/api/auth/[...nextauth]/route.ts
@@ -1,93 +1,5 @@
-import NextAuth, { DefaultSession, Account } from "next-auth";
-import { JWT } from "next-auth/jwt";
-import StravaProvider from "next-auth/providers/strava";
-
-const STRAVA_AUTH_URL = "https://www.strava.com/oauth/authorize?" + new URLSearchParams({
-  client_id: process.env.STRAVA_CLIENT_ID || '',
-  redirect_uri: `${process.env.NEXTAUTH_URL}/api/auth/callback/strava`,
-  response_type: 'code',
-  scope: 'read,activity:read_all',
-  approval_prompt: 'auto'
-}).toString();
-
-declare module "next-auth" {
-  interface Session extends DefaultSession {
-    accessToken?: string;
-    refreshToken?: string;
-    expiresAt?: number;
-  }
-}
-
-async function refreshAccessToken(token: JWT) {
-  try {
-    const response = await fetch('https://www.strava.com/oauth/token', {
-      method: 'POST',
-      headers: { 'Content-Type': 'application/x-www-form-urlencoded' },
-      body: new URLSearchParams({
-        client_id: process.env.STRAVA_CLIENT_ID!,
-        client_secret: process.env.STRAVA_CLIENT_SECRET!,
-        grant_type: 'refresh_token',
-        refresh_token: token.refreshToken as string,
-      }),
-    });
-
-    const refreshedTokens = await response.json();
-
-    if (!response.ok) {
-      throw refreshedTokens;
-    }
-
-    return {
-      ...token,
-      accessToken: refreshedTokens.access_token,
-      refreshToken: refreshedTokens.refresh_token ?? token.refreshToken,
-      expiresAt: refreshedTokens.expires_at,
-    };
-  } catch (error) {
-    console.error('Error refreshing access token', error);
-    return {
-      ...token,
-      error: "RefreshAccessTokenError",
-    };
-  }
-}
-
-export const authOptions = {
-  providers: [
-    StravaProvider({
-      clientId: process.env.STRAVA_CLIENT_ID!,
-      clientSecret: process.env.STRAVA_CLIENT_SECRET!,
-      authorization: STRAVA_AUTH_URL,
-    }),
-  ],
-  callbacks: {
-    async jwt({ token, account }: { token: JWT; account: Account | null }) {
-      if (account) {
-        // Initial sign in
-        return {
-          accessToken: account.access_token,
-          refreshToken: account.refresh_token,
-          expiresAt: account.expires_at,
-        };
-      }
-
-      // Return previous token if the access token has not expired yet
-      if (token.expiresAt && Date.now() < token.expiresAt * 1000) {
-        return token;
-      }
-
-      // Access token has expired, try to refresh it
-      return refreshAccessToken(token);
-    },
-    async session({ session, token }: { session: any; token: JWT }) {
-      session.accessToken = token.accessToken;
-      session.refreshToken = token.refreshToken;
-      session.expiresAt = token.expiresAt;
-      session.error = token.error;
-      return session;
-    },
-  },
-};
+import NextAuth from "next-auth";
+import { authOptions } from "./options";
 
 const handler = NextAuth(authOptions);
